Refactor TableView into separate component from Listing
This should enable us to easily swap in a card/thumbnail view and reuse functions from Listing in that view
