+++ src/components/data/Listing.js
@@ -1,33 +1,12 @@
 import React, { useState } from "react";
 
 import { useQuery } from "@apollo/react-hooks";
-import {
-    build,
-    DECheckbox,
-    DETableRow,
-    EmptyTable,
-    EnhancedTableHead,
-    formatDate,
-    getMessage,
-    LoadingMask,
-    withI18N,
-} from "@cyverse-de/ui-lib";
-import {
-    Paper,
-    Table,
-    TableBody,
-    TableCell,
-    TableContainer,
-    useMediaQuery,
-    useTheme,
-} from "@material-ui/core";
+import { LoadingMask, withI18N, } from "@cyverse-de/ui-lib";
+import { useMediaQuery, useTheme, } from "@material-ui/core";
 import gql from "graphql-tag";
 
-import DataDotMenu from "./DataDotMenu";
-import { getFileSize } from "./FileSize";
-import ids from "./ids";
 import messages from "./messages";
-import ResourceIcon from "./ResourceIcon";
+import TableView from "./TableView";
 
 const GET_LISTING = gql`
     query listing($path: String) {
@@ -53,41 +32,14 @@ const GET_LISTING = gql`
     }
 `;
 
-function getTableColumns(isMedium, isLarge) {
-    let columns = [
-        { name: "", align: "center", enableSorting: false, key: "icon" },
-        { name: "Name", align: "left", enableSorting: true },
-    ];
-
-    if (isMedium) {
-        columns.push(
-            {
-                name: "Last Modified",
-                align: "left",
-                enableSorting: false,
-                key: "creator",
-            },
-            { name: "Size", align: "left", enableSorting: false }
-        );
-    }
-
-    if (isLarge) {
-        columns.push({ name: "Path", align: "left", enableSorting: false });
-    }
-
-    columns.push({ name: "", align: "left", enableSorting: false });
-
-    return columns;
-}
 
 function Listing(props) {
     const theme = useTheme();
     let isMedium = useMediaQuery(theme.breakpoints.up("sm"));
     let isLarge = useMediaQuery(theme.breakpoints.up("lg"));
-    const tableColumns = getTableColumns(isMedium, isLarge);
 
     const [order, setOrder] = useState("asc");
-    const [orderBy, setOrderBy] = useState(tableColumns[0].name);
+    const [orderBy, setOrderBy] = useState("label");
     const [selected, setSelected] = useState([]);
     const [lastSelectIndex, setLastSelectIndex] = useState(-1);
     // const [page, setPage] = useState(0);
@@ -183,6 +135,22 @@ function Listing(props) {
         setSelected(newSelected);
     };
 
+    const onDownloadSelected = (resourceId) => {
+        console.log("Download", resourceId)
+    };
+
+    const onEditSelected = (resourceId) => {
+        console.log("Edit", resourceId)
+    };
+
+    const onMetadataSelected = (resourceId) => {
+        console.log("Metadata", resourceId)
+    };
+
+    const onDeleteSelected = (resourceId) => {
+        console.log("Delete", resourceId)
+    };
+
     // const handleChangePage = (event, newPage) => {
     //     setPage(newPage);
     // };
@@ -192,113 +160,25 @@ function Listing(props) {
     //     setPage(0);
     // };
 
-    let tableId = build(baseId, ids.listingTable);
-
     return (
         <LoadingMask loading={loading}>
-            <TableContainer component={Paper}>
-                <Table
-                    stickyHeader
-                    id={tableId}
-                    aria-label={getMessage("ariaTableListing")}
-                >
-                    <EnhancedTableHead
-                        selectable={true}
-                        numSelected={selected.length}
-                        onSelectAllClick={handleSelectAllClick}
-                        rowsInPage={listing?.length}
-                        order={order}
-                        orderBy={orderBy}
-                        baseId={tableId}
-                        columnData={tableColumns}
-                        onRequestSort={handleRequestSort}
-                    />
-                    <TableBody>
-                        {(!listing || listing.length === 0) && (
-                            <EmptyTable
-                                message={getMessage("emptyDataListing")}
-                                numColumns={tableColumns.length}
-                            />
-                        )}
-                        {listing &&
-                            listing.length > 0 &&
-                            listing.map((resource, index) => {
-                                const resourceId = resource.id;
-                                const isSelected =
-                                    selected.indexOf(resourceId) !== -1;
-                                return (
-                                    <DETableRow
-                                        role="checkbox"
-                                        tabIndex={-1}
-                                        hover
-                                        id={build(tableId, resourceId)}
-                                        key={resourceId}
-                                        selected={isSelected}
-                                        aria-checked={isSelected}
-                                        onClick={(event) =>
-                                            handleClick(
-                                                event,
-                                                resourceId,
-                                                index
-                                            )
-                                        }
-                                    >
-                                        <TableCell padding="checkbox">
-                                            <DECheckbox
-                                                checked={isSelected}
-                                                inputProps={{
-                                                    "aria-labelledby": build(
-                                                        tableId,
-                                                        resourceId
-                                                    ),
-                                                }}
-                                            />
-                                        </TableCell>
-                                        <TableCell padding="checkbox">
-                                            <ResourceIcon
-                                                type={resource.type}
-                                            />
-                                        </TableCell>
-                                        <TableCell>{resource.label}</TableCell>
-                                        {isMedium && (
-                                            <TableCell>
-                                                {formatDate(
-                                                    resource.dateModified,
-                                                    "YYYY MMM DD HH:mm:ss"
-                                                )}
-                                            </TableCell>
-                                        )}
-                                        {isMedium && (
-                                            <TableCell>
-                                                {getFileSize(resource.fileSize)}
-                                            </TableCell>
-                                        )}
-                                        {isLarge && (
-                                            <TableCell>{resource.path}</TableCell>
-                                        )}
-                                        <TableCell>
-                                            <DataDotMenu
-                                                baseId={tableId}
-                                                onDownloadSelected={() =>
-                                                    console.log("Download")
-                                                }
-                                                onEditSelected={() =>
-                                                    console.log("Edit")
-                                                }
-                                                onMetadataSeleted={() =>
-                                                    console.log("Metadata")
-                                                }
-                                                onDeleteSelected={() =>
-                                                    console.log("Delete")
-                                                }
-                                            />
-                                        </TableCell>
-                                    </DETableRow>
-                                );
-                            })}
-                    </TableBody>
-                </Table>
-            </TableContainer>
+            <TableView
+                path={path}
+                listing={listing}
+                isMedium={isMedium}
+                isLarge={isLarge}
+                baseId={baseId}
+                onDownloadSelected={onDownloadSelected}
+                onEditSelected={onEditSelected}
+                onMetadataSelected={onMetadataSelected}
+                onDeleteSelected={onDeleteSelected}
+                handleRequestSort={handleRequestSort}
+                handleSelectAllClick={handleSelectAllClick}
+                handleClick={handleClick}
+                order={order}
+                orderBy={orderBy}
+                selected={selected}
+            />
         </LoadingMask>
     );
 }
